+++ ohmg/frontend/svelte/src/lib/components/interfaces/buttons/FullExtentButton.svelte
@@ -4,8 +4,9 @@
     import CornersOut from "phosphor-svelte/lib/CornersOut";
 
     export let action;
+    export let style = "tool-ui";
 </script>
 
-<IconButton style="tool-ui" title="Go to full extent" action={action}>
+<IconButton {style} title="Go to full extent" action={action}>
     <CornersOut />
 </IconButton>
