+++ ohmg/frontend/svelte/src/lib/components/overviews/modals/MapPreviewModal.svelte
